Remove fs-verity build manifest for vendor/odm/product
The build manifest APKs are recently added because we thought it's nicer
for ART to support beyond system_ext for potential partner
customizations.  Jiyong pointed out that having system server JARs
outside of system and system_ext may be a Treble violation. As a result,
delete some of those files.

Bug: 246000387
Bug: 245957815
Test: m; no longer seeing the files under out/
Change-Id: I4c5b118bd2277b5f3265ac4d3c2343cf929822c5